+++ src/x86_64/include/font.h
@@ -15,6 +15,6 @@ typedef struct font {
 	const ch (*glyphs)[8];
 } font_t;
 
-kresult_t font_setup(font_t* font, const u32 width, const u32 height, const ch (*glyphs)[8]);
+kresult_t font_setup(font_t* font, const usize width, const usize height, const ch (*glyphs)[8]);
 const ch* font_get_glyph(const font_t* font, const ch c);
-b8 font_get_pixel_at(const font_t* font, const ch c, const u32 pixel_x, const u32 pixel_y);
+b8 font_get_pixel_at(const font_t* font, const ch c, const usize pixel_x, const usize pixel_y);
